[WiP] New idea: fold all selected types
Allow the user to chose exactly the type of nodes they want to fold
when folding all.

The current implementation expects the user to know the values in the
tree-sitter syntax tree which is not really user-friendly.

My idea would be to put this function behind an imenu or something
like this to show exactly what are the possible syntax values that can
be folded