Handle errors from treesit-buffer-root-node
It seems `treesit-buffer-root-node` returns an error when there is no
parser for a given buffer. In my case, this happens when trying to get
doc for an elisp variable; `ts-fold--tree-sitter-trigger` runs in an
elisp-mode "*temp*" buffer, and `treesit-buffer-root-node` raises an
error.
If I just ignore the error around that call,
`ts-fold--tree-sitter-trigger` still runs `(ts-fold-mode -1)`. Seems
like the simplest thing is to just do nothing in that case.

To repro the problem, enable ts-fold mode globally and do `C-H v
emacs-version`.