Optimize performance of ts-fold-close-all with indicators-mode on
Since ts-fold-close-all and ts-fold-close are both advised,
ts-fold-indicators-refresh will be called every time ts-fold-close-all calls
ts-fold-close on target nodes. Then ts-fold-close-all is far from being
responsive with indicators-mode on.
So we just remove the advice before we call ts-fold-close in ts-fold-close-all,
and get it back before we leave. Of course, ts-fold-indicators-refresh will be
called as an advice of ts-fold-close-all which we didn't modify.